Implement piped extractions. Extractions can be piped by adding the command arguments --all or --until. If --all is provided, the extractor will keep extracting until it reaches a format that it cannot handle. If --until is provided, the extractor will keep extracting until it reaches the extension specified by this parameter, or until it reaches a format it cannot handle.
Example. src: dir/Test.jar.pack.xz dest: dir/output
/ze extract --all will extract all the jar contents into output.
/ze extract --until will extract Test.jar into output.
/ze extract will extract Test.jar.pack into output (non piped).

If the piped operation contains a format which cannot be scanned ahead of time, the destination directory must be empty. If it is not, the operation will be cancelled, even if --override is passed. An example of this is the ZIP format. We cannot scan for extractions without first having the source file in the ZIP format. The scan involed checking each ZIP entry, which we do not have access to until the file is in ZIP format. This is the same for JAR and RAR.